+++ packages/cmf/src/cmfModule.merge.js
@@ -144,6 +144,6 @@ function merge(...configs) {
 
 export default merge;
 
-merge.mergeObjects = mergeObjects;
 merge.getReduceConfig = getReduceConfig;
 merge.getUnique = getUnique;
+merge.mergeObjects = mergeObjects;