+++ packages/containers/src/index.js
@@ -6,7 +6,7 @@ import * as containers from './containers';
 const components = Object.keys(allComponents).reduce((acc, key) => {
 	if (!acc[key] && typeof allComponents[key] === 'function') {
 		const options = {};
-		if (['ActionList', 'Layout', 'RichLayout', 'Dialog'].includes(key)) {
+		if (['ActionList', 'AppSwitcher', 'Layout', 'RichLayout', 'Dialog'].includes(key)) {
 			options.withComponentRegistry = true;
 		}
 		if (!allComponents[key].displayName) {