+++ packages/faceted-search/src/components/Badges/BadgeFaceted/BadgeFaceted.component.js
@@ -75,6 +75,8 @@ const BadgeFaceted = ({
 				{
 					value: badgeValue,
 					operator: badgeOperator,
+					initialOperatorOpened: false,
+					initialValueOpened: false,
 				},
 				{ isInCreation: false },
 			),