+++ packages/faceted-search/CHANGELOG.md
@@ -26,6 +26,8 @@ Types of changes
 
 ## [unreleased]
 
+## [0.13.0]
+
 - [Fixed](https://github.com/Talend/ui/pull/2972): Don't reopen the value tooltip for badges with default operators
 - [Added](https://github.com/Talend/ui/pull/2932): Pointer cursor on the slider facet value